REFACTOR: move getOverdueTasks function users model to tasks model
